app_logic/requirements.txt
langchain-google-genai
langchain
langchain-community
python-dotenv
streamlit
wikipedia
langchain-classic    


langchain-ollama
langchain-chroma
pandas

langchain_text_splitters
pypdf
docx2txt
python-pptx
pytesseract
Pillow

chromadb
google-genai   
openai